ObjectIndex: support PluginIndex better
We used to require overriding the add() method for only one reason:
the type by which the objects are indexed in PluginIndex are not
obtained by getClass(), but by getPluginType().

This suggests that the base class actually needs a more flexible way
to obtain the types by which to index, so let's do that instead.

This will help with the upcoming performance enhancements which will
bypass the add() method completely for SortedObjectIndex' addAll().

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>