Try to speed up the SortedObjectIndex further
Rather than adding things blindly to several typed lists which are all
supported to be sorted already, only to sort them again in the end,
let's

- sort the elements to be added

- make sub-lists of the sorted list for the individual typed lists

- merge the sorted lists in O(n)

instead.

While at it, we should not iterate the elements of HashSets with
supposedly immutable contents over and over and over again. Let's make
performant arrays from them as soon as we generated the sets.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>